requirements.txt
astroid==2.8.0
beautifulsoup4==4.10.0
blurhash==1.1.4
certifi==2021.5.30
charset-normalizer==2.0.5
decorator==5.1.0
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
Mastodon.py==1.5.1
mccabe==0.6.1
pkg-resources==0.0.0
platformdirs==2.3.0
pylint==2.11.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-magic==0.4.24
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
schedule==1.1.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.6
wrapt==1.12.1
